referencing

Struct Resolver

pub struct Resolver<'r> { /* private fields */ }

A reference resolver.

Resolves references against the base URI and looks up the result in the registry.

Implementations

impl<'r> Resolver<'r>

pub fn base_uri(&self) -> Arc<Uri<String>>

Returns the current base URI of this resolver.

pub fn lookup(&self, reference: &str) -> Result<Resolved<'r>, Error>

Resolve a reference to the resource it points to.

Errors

If the reference cannot be resolved or is invalid.

pub fn lookup_recursive_ref(&self) -> Result<Resolved<'r>, Error>

Resolve a recursive reference.

This method implements the recursive reference resolution algorithm as specified in JSON Schema Draft 2019-09.

It starts by resolving “#” and then follows the dynamic scope, looking for resources with $recursiveAnchor: true.

Errors

This method can return any error that Resolver::lookup can return.

pub fn in_subresource( &self, subresource: ResourceRef<'_>, ) -> Result<Self, Error>

Create a resolver for a subresource.

Errors

Returns an error if the resource id cannot be resolved against the base URI of this resolver.

pub fn dynamic_scope(&self) -> List<Uri<String>>

pub fn resolve_uri( &self, base: &Uri<&str>, uri: &str, ) -> Result<Arc<Uri<String>>, Error>

Resolve a reference against a base.

Errors

If the reference is invalid.

pub fn find_vocabularies(&self, draft: Draft, contents: &Value) -> VocabularySet